Added recommender-not working

+++ backend/src/main/java/com/esaunders/library/dtos/BookDto.java
@@ -9,7 +9,7 @@ public class BookDto {
     private String genre;
     private String thumbnail;
     private String description;
-    private double year;
+    private int year;
     private double rating;
     private int pages;
 }
